consolidate tests into one file for now
